+++ qbraid/visualization/__init__.py
@@ -23,6 +23,7 @@ Functions
     plot_histogram
     plot_distribution
     plot_conversion_graph
+    plot_runtime_conversion_scheme
     circuit_drawer
     qasm3_drawer
     plot_atomic_register
@@ -42,13 +43,14 @@ from .draw_circuit import circuit_drawer
 from .draw_qasm3 import qasm3_drawer
 from .exceptions import VisualizationError
 from .flair_animations import animate_qpu_state
-from .plot_conversions import plot_conversion_graph
+from .plot_conversions import plot_conversion_graph, plot_runtime_conversion_scheme
 from .plot_counts import plot_distribution, plot_histogram
 
 __all__ = [
     "plot_histogram",
     "plot_distribution",
     "plot_conversion_graph",
+    "plot_runtime_conversion_scheme",
     "circuit_drawer",
     "qasm3_drawer",
     "plot_atomic_register",